# study-service/src/main/resources/openapi/study-api.openapi.yml
openapi: 3.0.0
info:
  title: Study API
  version: 0.0.1
servers:
  - url: http://localhost:8083/api/study
paths:
  /schedule:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get All Schedule
      tags:
        - Schedule
      operationId: getAllSchedule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '400':
          description: Bad Request
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Schedule
      tags:
        - Schedule
      operationId: createSchedule
      requestBody:
        description: Schedule Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Schedule by Id
      tags:
        - Schedule
      operationId: updateSchedule
      requestBody:
        description: Schedule Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleOpenAPI'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /schedule/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Schedule Information by Id
      tags:
        - Schedule
      operationId: getScheduleByGroup
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Schedule by Id
      tags:
        - Schedule
      operationId: deleteSchedule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /schedule/professor/{professorId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Schedule By Professor Id
      tags:
        - Schedule
      operationId: getScheduleByProfessorId
      parameters:
        - in: path
          name: professorId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '400':
          description: Bad Request
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

  /group/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group Information
      tags:
        - Group
      operationId: getGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Group By Id
      tags:
        - Group
      operationId: deleteGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /group:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Groups Information
      tags:
        - Group
      operationId: getAllGroups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Group
      tags:
        - Group
      operationId: createGroup
      requestBody:
        description: Group Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /materials:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Materials Information
      tags:
        - Materials
      operationId: getAllMaterials
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Material
      tags:
        - Materials
      operationId: createMaterials
      requestBody:
        description: Materials Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyMaterialsOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Materials
      tags:
        - Materials
      operationId: updateMaterials
      requestBody:
        description: Materials Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyMaterialsOpenAPI'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /materials/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Materials Information
      tags:
        - Materials
      operationId: getMaterialsById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Material by Id
      tags:
        - Materials
      operationId: deleteMaterials
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

  /materials/{materialsId}/homework/all:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Homeworks Information by Materials Id
      tags:
        - Homework
      operationId: getAllHomeworksByMaterials
      parameters:
        - in: path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

  /materials/{materialsId}/homework:
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Homework
      tags:
        - Homework
      operationId: createHomework
      parameters:
        - in:
            path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Homework Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaveHomeworkRequestOpenApi'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Homework
      tags:
        - Homework
      operationId: updateHomework
      parameters:
        - in: path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Homework Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaveHomeworkRequestOpenApi'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /materials/{materialsId}/homework/{homeworkId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Homework Information
      tags:
        - Homework
      operationId: getHomeworkById
      parameters:
        - in: path
          name: materialsId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: homeworkId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Homework
      tags:
        - Homework
      operationId: deleteHomework
      parameters:
        - in: path
          name: materialsId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: homeworkId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /homework/student/{studentId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Homeworks Information By Student
      tags:
        - Homework
      operationId: getHomeworkByStudentId
      parameters:
        - name: studentId
          in: path
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Student Not Found

  /materials/{materialsId}/homework/student/{studentId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Student's Homework by Material
      tags:
        - Homework
      operationId: getHomeworkByStudentAndMaterials
      parameters:
          - name: materialsId
            in: path
            schema:
              type: integer
              format: int64
            required: true
          - name: studentId
            in: path
            schema:
              type: integer
              format: int64
            required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Student Not Found
    put:
      security:
        - bearerAuth: [ ]
      summary: Grade Student's Homework
      tags:
        - Homework
      operationId: gradeHomeworkByStudent
      parameters:
        - name: materialsId
          in: path
          schema:
            type: integer
            format: int64
          required: true
        - name: studentId
          in: path
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Task Evaluation Object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskEvaluationOpenAPI"
      responses:
        '200':
          description: OK
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Student Not Found

  /subject:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Subjects Information
      tags:
        - Subject
      operationId: getAllSubjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Subject
      tags:
        - Subject
      operationId: createSubject
      requestBody:
        description: Subject Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Subject
      tags:
        - Subject
      operationId: updateSubject
      requestBody:
        description: Subject Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectOpenAPI'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

  /subject/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject Information
      tags:
        - Subject
      operationId: getSubjectById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Subject by Id
      tags:
        - Subject
      operationId: deleteSubject
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found


components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    ScheduleOpenAPI:
      type: object
      required:
        - dateTime
        - studyGroup
        - subject
        - classFormat
        - professor
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        studyGroup:
          $ref: "#/components/schemas/StudyGroupOpenAPI"
        subject:
          $ref: "#/components/schemas/SubjectOpenAPI"
        classFormat:
          $ref: "#/components/schemas/ClassFormat"
        auditorium:
          type: string
        linkForTheClass:
          type: string
        professor:
          $ref: "#/components/schemas/ProfessorSubjectOpenAPI"

    StudyGroupOpenAPI:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    SubjectOpenAPI:
      type: object
      required:
        - name
        - studyGroup
        - creatorId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studyGroup:
          type: array
          items:
            $ref: "#/components/schemas/StudyGroupOpenAPI"
        studyMaterials:
          type: array
          items:
            $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleOpenAPI"
        creatorId:
          type: integer
          format: int64
        editorsIds:
          type: array
          items:
            type: integer
            format: int64

    ProfessorSubjectOpenAPI:
      type: object
      required:
        - subjectId
        - professorId
      properties:
        subjectId:
          type: integer
          format: int64
        professorId:
          type: integer
          format: int64

    ClassFormat:
      type: string
      enum:
        - IN_PERSON
        - DISTANT

    StudyMaterialsType:
      type: string
      enum:
        - MATERIAL
        - TASK

    StudyMaterialsOpenAPI:
      type: object
      required:
        - studyMaterialsType
        - subject
        - materialsText
      properties:
        id:
          type: integer
          format: int64
        studyMaterialsType:
          $ref: "#/components/schemas/StudyMaterialsType"
        materialsText:
          type: string
        subject:
          $ref: "#/components/schemas/SubjectOpenAPI"
        homework:
          type: array
          items:
            $ref: "#/components/schemas/HomeworkOpenAPI"

    HomeworkOpenAPI:
      type: object
      required:
        - task
        - studentId
        - date
      properties:
        id:
          type: integer
          format: int64
        task:
          $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        studentId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        grade:
          type: integer
        comment:
          type: string

    SaveHomeworkRequestOpenApi:
      type: object
      required:
        - homework
        - file
      properties:
        homework:
          $ref: "#/components/schemas/HomeworkOpenAPI"
        file:
          type: array
          items:
            type: string
            format: binary
            description: File with homework

    TaskEvaluationOpenAPI:
      type: object
      required:
        - grade
      properties:
        grade:
          type: integer
        comment:
          type: string